Refactoring the TrickTypeHandler.
